% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilisk.R
\docType{data}
\name{envPirat}
\alias{envPirat}
\title{Creates a BasiliskEnvironment class}
\format{
An object of class \code{BasiliskEnvironment} of length 1.
}
\usage{
envPirat
}
\value{
An instance of the class `BasiliskEnvironment`
}
\description{
Please refer to the package `basilisk`.
}
\keyword{datasets}
