% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_impute.R
\name{estimate_gamma}
\alias{estimate_gamma}
\title{Estimate missingness parameters Gamma}
\usage{
estimate_gamma(pep.ab.table, mcar = FALSE)
}
\arguments{
\item{pep.ab.table}{The peptide or precrursor abundance matrix, with 
molecules in columns and samples in row.}

\item{mcar}{If TRUE, forces gamma_1 = 0.}
}
\value{
A list of the containing missingness parameters gamma_0 and gamma_1.
}
\description{
Estimate missingness parameters Gamma
}
\examples{
data(subbouyssie)
estimate_gamma(subbouyssie$peptides_ab)


}
