% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirat_utils.R
\name{get_indexes_embedded_prots}
\alias{get_indexes_embedded_prots}
\title{Indexes of PGs embedded in each others}
\usage{
get_indexes_embedded_prots(adj)
}
\arguments{
\item{adj}{An adjacency matrix between precursors/peptides and PGs}
}
\value{
A vector of indices
}
\description{
Returns indexes of PGs that are embedded in others
}
\examples{
data(subbouyssie)
get_indexes_embedded_prots(subbouyssie$adj)

}
