% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_solo_pep}
\alias{group_solo_pep}
\title{xxx}
\usage{
group_solo_pep(adj, group_size)
}
\arguments{
\item{adj}{xxx}

\item{group_size}{xxx}
}
\value{
xxx
}
\description{
Return adjacency matrix where peptides that are alone in their component are grouped
}
\examples{
NULL

}
