% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_from_blocks}
\alias{impute_from_blocks}
\title{xxx}
\usage{
impute_from_blocks(logs.blocks, data.pep.rna, idx_blocks)
}
\arguments{
\item{logs.blocks}{xxx}

\item{data.pep.rna}{xxx}

\item{idx_blocks}{xxx}
}
\value{
xxx
}
\description{
Returns the imputed matrix from the imputed PGs.
}
\examples{
NULL

}
