% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_Pirat2.R
\name{install_pirat}
\alias{install_pirat}
\title{Install Pirat package}
\usage{
install_pirat()
}
\description{
This script installs Python and PyTorch in the requested
versions. It is largely inspired by wthe scripts in rTorch package
(https://github.com/f0nzie/rTorch)
}
\examples{
\dontrun{

# install PyTorch 1.10.0 on Python 3.3.9.5 including pandas
install_pytorch(version = "1.10.0", conda_python_version = "3.9.5",
extra_packages = "pandas")

# Install PyTorch 1.10.0, Python 3.9.5, pandas, matplotlib install from the console
install_pytorch(version = "1.10.0", conda_python_version = "3.9.5",
extra_packages = c("pandas", "matplotlib"))

# Install PyTorch 1.10.0 on Python 3.9.5 including pandas, matplotlib
install_pytorch(version = "1.10.0", conda_python_version = "3.9.5",
extra_packages = c("pandas", "matplotlib"), dry_run = FALSE)
}

}
