% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_Pirat_env}
\alias{install_Pirat_env}
\title{Install python environment for Pirat}
\usage{
install_Pirat_env()
}
\value{
A list of names of objects exposed in each module.
}
\description{
This function may be run only once and use the package
`basilisk` to install the Python environment needed to run the 
imputation functions in the package `Pirat`
}
\examples{
install_Pirat_env()

}
\author{
Samuel Wieczorek
}
