% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Pirat.R
\name{mod_Pirat}
\alias{mod_Pirat}
\alias{mod_Pirat_ui}
\alias{mod_Pirat_server}
\title{## To be customized ##}
\usage{
mod_Pirat_ui(id)

mod_Pirat_server(
  id,
  obj = reactive({
     NULL
 }),
  reset = reactive({
     NULL
 }),
  verbose = FALSE
)
}
\arguments{
\item{id}{The id of the module}

\item{obj}{An list of two items}

\item{reset}{xxx}

\item{verbose}{A boolean (FALSE as default) which indicates whether to 
display more details on the process}
}
\value{
A shiny app

A shiny app

A shiny app
}
\description{
## To be customized ##
}
