% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{Pirat}
\alias{Pirat}
\alias{Pirat-package}
\title{xxx}
\value{
NA
}
\description{
xxxx
# https://community.rstudio.com/t/when-to-use-onload-vs-onattach/21953
Usually you want .onLoad, which—as the name suggests—runs when the package is 
loaded. If something has to happen before anything is run, that's the way to 
go. onAttach only runs when the library is attached, e.g. when somebody calls
library(your_package). onLoad will also run when somebody loads but doesn't 
attach your package by calling your_package::your_function.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.prostar-proteomics.org/}
  \item Report bugs at \url{https://github.com/prostarproteomics/Pirat/issues}
}

}
\author{
\strong{Maintainer}: Samuel Wieczorek \email{samuel.wieczorek@cea.fr}

Authors:
\itemize{
  \item Lucas Etourneau \email{lucas.etourneau@univ-grenoble-alpes.fr}
  \item Laura Fancello \email{laura.fancello@cea.fr}
  \item Nelle Varoquaux \email{nelle.varoquaux@univ-grenoble-alpes.fr}
  \item Thomas Burger \email{thomas.burger@cea.fr}
}

}
