% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{pirat2SE}
\alias{pirat2SE}
\title{COnvert Pirat dataset to SummarizedExperiment}
\usage{
pirat2SE(peptides_ab, adj, mask_prot_diff, mask_pep_diff)
}
\arguments{
\item{peptides_ab}{xxx}

\item{adj}{xxx}

\item{mask_prot_diff}{xxx}

\item{mask_pep_diff}{xxx}
}
\value{
An instance of the class `SummarizedExperiment`
}
\description{
This function converts the original dataset structure into 
a SummarizedExperiment .
}
\examples{
data(subbouyssie)
peptides_ab <- subbouyssie$peptides_ab
adj <- subbouyssie$adj
mask_prot_diff <- subbouyssie$mask_prot_diff
mask_pep_diff <- subbouyssie$mask_pep_diff
obj <- pirat2SE(peptides_ab, adj, mask_prot_diff, mask_pep_diff )
obj

}
