% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pirat_config}
\alias{pirat_config}
\alias{torch_version}
\title{Pirat configuration information}
\usage{
pirat_config()

torch_version()
}
\value{
List with information on the current configuration of TensorFlow.
  You can determine whether TensorFlow was found using the `available`
  member (other members vary depending on whether `available` is `TRUE`
  or `FALSE`)
}
\description{
Pirat configuration information
}
\keyword{internal}
