% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot2hists}
\alias{plot2hists}
\title{xxx}
\usage{
plot2hists(d1, d2, name1, name2, titlename, xlab = "", freq = T)
}
\arguments{
\item{d1}{xxx}

\item{d2}{xxx}

\item{name1}{xxx}

\item{name2}{xxx}

\item{titlename}{xxx}

\item{xlab}{xxx}

\item{freq}{xxx}
}
\description{
xxxxx
}
