% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pep_correlations}
\alias{plot_pep_correlations}
\title{Empirical density of peptide correlations}
\usage{
plot_pep_correlations(pep.data, titlename = NULL, xlabel = "Correlations")
}
\arguments{
\item{pep.data}{List representing dataset}

\item{titlename}{Title of the graph displayed}

\item{xlabel}{Label of x-axis}
}
\value{
The ggplot2 graph
}
\description{
Plot empirical densities of correlations between peptides within
 PG and at random, estimated by gaussian kernel. Note that only correlations 
 between fully observed peptides are considered here.
}
\examples{
data(subbouyssie)
plot_pep_correlations(subbouyssie, 'test')

}
