% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pirat_utils.R
\name{Py_impute_block_llk_reset}
\alias{Py_impute_block_llk_reset}
\title{Wrapper for the function `impute_block_llk_reset()`}
\usage{
Py_impute_block_llk_reset(data.pep.rna.mis, psi)
}
\arguments{
\item{data.pep.rna.mis}{xxx}

\item{psi}{Inverse scale parameter for IW prior of peptides abundances}
}
\value{
The original peptide abundance table with imputed values.
}
\description{
Wrapper for the function `impute_block_llk_reset()`
}
\examples{
\dontrun{
#' data(subbouyssie)
obj <- subbouyssie
# Keep only fully observed peptides
obs2NApep <- obj$peptides_ab[ ,colSums(is.na(obj$peptides_ab)) <= 0] 
res_hyperparam <- estimate_psi_df(obs2NApep)
psi <- res_hyperparam$psi
imputed_pgs <- Py_impute_block_llk_reset(obj, psi)
}
}
