% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_pep_from_idx}
\alias{remove_pep_from_idx}
\title{Remove peptides by index}
\usage{
remove_pep_from_idx(l_pep_rna, pepidxs)
}
\arguments{
\item{l_pep_rna}{List representing the dataset}

\item{pepidxs}{Vector of indices}
}
\value{
List representing the dataset
}
\description{
Remove peptide by index, then deletes empty PGs
}
