% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{requested_versions}
\alias{requested_versions}
\title{List of versions for Python packages needed by the package Pirat}
\format{
An object of class \code{list} of length 3.
}
\usage{
requested_versions
}
\value{
A list
}
\description{
List of versions for Python packages needed by the package Pirat
}
\keyword{datasets}
