% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{ropers}
\alias{ropers}
\title{Ropers dataset}
\format{
A list containing:
- peptides_ab: numeric matrix of precrusors log2 abundances.
- adj: adjacency matrix between peptides and PGs
- rnas_ab: numeric matrix of gene expression log2 counts from mRNA analysis.
- adj_rna_pg: adjacency matrix between genes and PGs
- charges: charge of each precursor.
- modifs: post-translational modification of each precursor.
}
\value{
A dataset
}
\description{
This dataset corresponds to `Ropers2021` dataset, described in 
Pirat article.
}
\references{
Ropers, D., Couté, Y., Faure, L., Ferré, S., Labourdette, D., 
Shabani, A., Trouilh, L., Vasseur, P., Corre, G., Ferro, M., Teste, M. A., 
Geiselmann, J., & de Jong, H. (2021). Multiomics Study of Bacterial Growth 
Arrest in a Synthetic Biology Application. ACS Synthetic Biology, 10(11), 
2910–2926. 
https://doi.org/10.1021/ACSSYNBIO.1C00115/SUPPL_FILE/SB1C00115_SI_010.ZIP
}
\keyword{data}
\keyword{datasets}
