% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_own_pyScripts}
\alias{source_own_pyScripts}
\title{Source user defined Python code}
\usage{
source_own_pyScripts()
}
\value{
NA
}
\description{
Use the function `reticulate::source_python()` to source
custom python files stored int the directory 'inst/python' of the 
package `Pirat`.
}
