% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_large_pg}
\alias{split_large_pg}
\title{xxx}
\usage{
split_large_pg(adj, size_max)
}
\arguments{
\item{adj}{xxx}

\item{size_max}{xxx}
}
\value{
xxx
}
\description{
Split PGs with too many peptides/precursors.
}
\examples{
NULL

}
