% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_funcs.R
\name{split_large_pg_PG}
\alias{split_large_pg_PG}
\title{Splits too large PGs in proteogenomics context}
\usage{
split_large_pg_PG(adj, size_max, adj_rna_pg)
}
\arguments{
\item{adj}{Adjacency matrix between peptides and PGs.}

\item{size_max}{Maximum PG size desired.}

\item{adj_rna_pg}{Adjacency matrix between mRNA and PGs.}
}
\value{
List containing new adjacency matrix between peptides and PGs, 
and new adjacency matrix between mRNA and PGs.
}
\description{
Split PGs with too many peptides/precursors while keeping other PGs untouched,
and adapts adjacency matrix between mRNA and PGs accordingly. It creates new PGs 
with size equal to size max (including peptides and mRNAs). Hence, some 
peptides and mRNA can be duplicated in the new PGs.
The seed is fixed such that the split generated remains constant for a given
adjacency matrix.
}
\examples{
NULL

}
