% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-data.R
\docType{data}
\name{subropers}
\alias{subropers}
\title{Sub-Ropers dataset}
\format{
A list of six items:
* peptides_ab: the peptide or precursor abundance matrix to impute, 
with samples in row and peptides or precursors in column
* rnas_ab: a n_peptide x n_protein adjacency matrix between peptides and 
proteins containing 0 and 1, or TRUE and FALSE.
* adj: A matrix which is the adjacency matrix for the slot peptides_ab.
* adj_rna_pg: a n_mrna x n_protein adjacency matrix n_mrna and proteins 
containing 0 and 1, or TRUE and FALSE
* charges: xxx
* modifs: xxx
}
\value{
A dataset
}
\description{
This dataset is extracted from the original `Ropers2021`
dataset described in Pirat article, where only 10 PGs were randomly selected.
}
\references{
Ropers, D., Couté, Y., Faure, L., Ferré, S., Labourdette, D., 
Shabani, A., Trouilh, L., Vasseur, P., Corre, G., Ferro, M., Teste, M. A., 
Geiselmann, J., & de Jong, H. (2021). Multiomics Study of Bacterial Growth 
Arrest in a Synthetic Biology Application. ACS Synthetic Biology, 10(11), 
2910–2926. 
https://doi.org/10.1021/ACSSYNBIO.1C00115/SUPPL_FILE/SB1C00115_SI_010.ZIP
}
\keyword{data}
\keyword{datasets}
