% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{wrapper_pipeline_llkimpute}
\alias{wrapper_pipeline_llkimpute}
\title{xxx}
\usage{
wrapper_pipeline_llkimpute(se, ...)
}
\arguments{
\item{se}{xxx}

\item{...}{xxx}
}
\description{
xxx
}
\examples{
data(bouyssie)

}
