% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_impute.R
\name{estimate_gamma}
\alias{estimate_gamma}
\title{xxx}
\usage{
estimate_gamma(pep.ab.table, mcar = FALSE)
}
\arguments{
\item{pep.ab.table}{xxx}

\item{mcar}{xxx}
}
\value{
A list
}
\description{
xxx
}
\examples{
NULL

}
