% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_impute.R
\name{estmate_psi_df}
\alias{estmate_psi_df}
\title{Estimate psi and degrees of freedom}
\usage{
estmate_psi_df(obs2NApep)
}
\arguments{
\item{obs2NApep}{Peptide abundance matrix (can contain MVs)}
}
\value{
List containing estimated parameters df (degrees of freedom) and psi (inverse scale)
}
\description{
Estimate the inverse scale factor and degrees of freedom of 
the distribution of columns-wise variances of an abundance table
}
\examples{
NULL

}
