% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplot2hist}
\alias{ggplot2hist}
\title{xxx}
\usage{
ggplot2hist(list.values, titlename, xlabel = "")
}
\arguments{
\item{list.values}{xxx}

\item{titlename}{xxx}

\item{xlabel}{xxx}
}
\value{
xxx
}
\description{
xxxx
}
\examples{
NULL

}
