% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplot2hist}
\alias{ggplot2hist}
\title{Plot several densities}
\usage{
ggplot2hist(list.values, titlename, xlabel = "")
}
\arguments{
\item{list.values}{xxx}

\item{titlename}{xxx}

\item{xlabel}{xxx}
}
\value{
xxx
}
\description{
Plot empirical densities estimated by Gaussian kernel
}
\examples{
ll <- list(A=1:10, B=3:6)
ggplot2hist(ll, 'test')

}
