% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_block_llk_reset}
\alias{impute_block_llk_reset}
\title{xxx}
\usage{
impute_block_llk_reset(
  data.pep.rna.crop,
  impfunc,
  psi,
  pep_ab_or = NULL,
  psiratio = 0,
  K_fixed = NULL,
  tol.na.pep = 100,
  prot.idxs = NULL,
  df = 1,
  nu_factor = 1,
  psi_rna = NULL,
  rna.cond.mask = F,
  group_pep_solo = F,
  max_pg_size = NULL,
  ...
)
}
\arguments{
\item{data.pep.rna.crop}{A list representing dataset}

\item{impfunc}{xxx}

\item{psi}{xxx}

\item{pep_ab_or}{xxx}

\item{psiratio}{xxx}

\item{K_fixed}{xxx}

\item{tol.na.pep}{xxx}

\item{prot.idxs}{xxx}

\item{df}{xxx}

\item{nu_factor}{xxx}

\item{psi_rna}{xxx}

\item{rna.cond.mask}{xxx}

\item{group_pep_solo}{xxx}

\item{max_pg_size}{xxx}

\item{...}{xxx}
}
\value{
xxx
}
\description{
Imputes dataset and return result for each component.
}
\examples{
NULL

}
