% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_block_llk_reset_PG}
\alias{impute_block_llk_reset_PG}
\title{xxx}
\usage{
impute_block_llk_reset_PG(
  data.pep.rna.crop,
  impfunc,
  psi,
  psi_rna,
  rna.cond.mask,
  pep.cond.mask,
  pep_ab_or = NULL,
  psiratio = 0,
  prot.idxs = NULL,
  df = 1,
  nu_factor = 1,
  max_pg_size = NULL,
  ...
)
}
\arguments{
\item{data.pep.rna.crop}{xxx}

\item{impfunc}{xxx}

\item{psi}{xxx}

\item{psi_rna}{xxx}

\item{rna.cond.mask}{xxx}

\item{pep.cond.mask}{xxx}

\item{pep_ab_or}{xxx}

\item{psiratio}{xxx}

\item{prot.idxs}{xxx}

\item{df}{xxx}

\item{nu_factor}{xxx}

\item{max_pg_size}{xxx}

\item{...}{xxx}
}
\value{
xxx
}
\description{
Impute dataset and return result for each component accounting for transcriptomic information
}
\examples{
NULL

}
