% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_Pirat.R
\name{install_pirat}
\alias{install_pirat}
\title{Install the package Pirat}
\usage{
install_pirat(force = FALSE)
}
\arguments{
\item{force}{A boolean (default value is FALSE) indicating whether to erase 
a current installation of the package Pirat.}
}
\description{
This script installs Python and PyTorch in their requested
versions. TSo as to be compliant with different OS, it only uses a conda
environment. This script is largely inspired by the scripts in the packages 
rTorch (https://github.com/f0nzie/rTorch) and tensorflow 
(https://github.com/rstudio/tensorflow).
}
\examples{
\dontrun{
install_pirat()
}

}
\author{
Samuel Wieczorek
}
