% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pirat.R
\name{install_pirat}
\alias{install_pirat}
\title{Install Python environment and libraries}
\usage{
install_pirat()
}
\value{
NA
}
\description{
This function installs the Python environment and the
Python libraries needed to run the imputation functions of the package 
`Pirat`. This operation takes several minutes.
It has to be execute only once, before the first use of the package.
}
\examples{
\donttest{
install_pirat()
}

}
