% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_Pirat.R
\name{install_Pirat}
\alias{install_Pirat}
\title{Install Pirat and its dependencies}
\usage{
install_Pirat(
  envname = "r-pirat",
  new_env = identical(envname, "r-pirat"),
  restart_session = TRUE
)
}
\arguments{
\item{envname}{xxx}

\item{new_env}{If \code{TRUE}, any existing Python virtual environment and/or
conda environment specified by \code{envname} is deleted first.}

\item{restart_session}{Whether to restart R session after installing. Note that
it will be automatically if in RStudio.}
}
\description{
\code{install_Pirat()} installs just the python packages needed by Pirat.
}
\details{
You may be prompted to download and install miniconda if reticulate
did not find a non-system installation of python. Miniconda is the
recommended installation method for most users, as it ensures that the R
python installation is isolated from other python installations. All python
packages will by default be installed into a self-contained conda or venv
environment named "r-Pirat". Note that "conda" is the only
supported method on M1 Mac.

If you initially declined the miniconda installation prompt, you can later
manually install miniconda by running \code{\link[reticulate:install_miniconda]{reticulate::install_miniconda()}}.
}
\section{Custom Installation}{
 \code{install_Pirat()} isn't required to use Pirat.
If you manually configure a python environment with the required
dependencies and Python environment, you can tell R to use it by pointing
reticulate at it, commonly by setting an environment variable:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{Sys.setenv("RETICULATE_PYTHON" = "~/path/to/python-env/bin/python")
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
install_Pirat()
}


}
