% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_Pirat2.R
\name{install_pirat}
\alias{install_pirat}
\title{Install Pirat package}
\usage{
install_pirat(
  method = c("auto", "virtualenv", "conda"),
  conda = "auto",
  envname = "r-pirat",
  restart_session = TRUE,
  channel = c("pytorch", "stable"),
  pip_ignore_installed = FALSE,
  new_env = identical(envname, "r-pirat")
)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method.  Note that since this command runs without
privilege the "system" method is available only on _Windows_.}

\item{conda}{www}

\item{restart_session}{Restart R session after installing (note this will
only occur within RStudio).}

\item{channel}{conda channel. The default channel is `stable`.
The alternative channel is `nightly`.}

\item{extra_packages}{Additional Python packages to install along with
PyTorch. Default are `c("numpy=1.20.2", "matplotlib")`.}

\item{pip}{logical}

\item{cuda_version}{string for the cuda toolkit version to install. For example,
to install a specific CUDA version use `cuda_version="10.2"`.}

\item{dry_run}{logical, set to TRUE for unit tests, otherwise will execute
the command.}

\item{...}{other arguments passed to [reticulate::conda_install()] or
[reticulate::virtualenv_install()].}
}
\description{
This script installs Python and PyTorch in the requested
versions. It is largely inspired by wthe scripts in rTorch package
(https://github.com/f0nzie/rTorch)
}
\examples{
\dontrun{

# install PyTorch 1.10.0 on Python 3.3.9.5 including pandas
install_pytorch(version = "1.10.0", conda_python_version = "3.9.5",
extra_packages = "pandas")

# Install PyTorch 1.10.0, Python 3.9.5, pandas, matplotlib install from the console
install_pytorch(version = "1.10.0", conda_python_version = "3.9.5",
extra_packages = c("pandas", "matplotlib"))

# Install PyTorch 1.10.0 on Python 3.9.5 including pandas, matplotlib
install_pytorch(version = "1.10.0", conda_python_version = "3.9.5",
extra_packages = c("pandas", "matplotlib"), dry_run = FALSE)
}

}
