% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_Pirat}
\alias{install_Pirat}
\title{Test function}
\usage{
install_Pirat()
}
\value{
A list of names of objects exposed in each module.
}
\description{
Does nothing but test that we can load modules from different virtual environments.
}
\examples{
install_Pirat()
}
\author{
Aaron Lun
}
