% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Pirat.R
\name{mod_Pirat}
\alias{mod_Pirat}
\alias{mod_Pirat_ui}
\alias{mod_Pirat_server}
\title{## To be customized ##}
\usage{
mod_Pirat_ui(id)

mod_Pirat_server(
  id,
  obj = reactive({
     NULL
 }),
  reset = reactive({
     NULL
 }),
  verbose = FALSE
)
}
\arguments{
\item{id}{xxx}

\item{obj}{xxx}

\item{reset}{xxx}

\item{verbose}{xxx}
}
\description{
## To be customized ##
}
