% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_impute.R
\name{pipeline_llkimpute}
\alias{pipeline_llkimpute}
\title{xxxx}
\usage{
pipeline_llkimpute(
  data.pep.rna.mis,
  pep.ab.comp = NULL,
  nu_factor = 2,
  cov_ratio = 0,
  rna.cond.mask = NULL,
  pep.cond.mask = NULL,
  group_pep_solo = FALSE,
  mcar = FALSE,
  protidxs = NULL,
  max.pg.size2imp = NULL,
  transpose = FALSE,
  degenerated = FALSE
)
}
\arguments{
\item{data.pep.rna.mis}{xxx}

\item{pep.ab.comp}{xxx}

\item{nu_factor}{xxx}

\item{cov_ratio}{xxx}

\item{rna.cond.mask}{xxx}

\item{pep.cond.mask}{xxx}

\item{group_pep_solo}{xxx}

\item{mcar}{xxx}

\item{protidxs}{xxx}

\item{max.pg.size2imp}{xxx}

\item{transpose}{xxx}

\item{degenerated}{xxx}
}
\value{
xxx
}
\description{
Pirat imputation function
}
\examples{
data(bouyssie)
res.mle <- pipeline_llkimpute(bouyssie)
res.mle.transp <- pipeline_llkimpute(bouyssie, transpose = TRUE)
res.mle.transp.mcar <- pipeline_llkimpute(bouyssie, transpose = TRUE, mcar = TRUE)

}
