% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_impute.R
\name{pipeline_llkimpute}
\alias{pipeline_llkimpute}
\title{Pirat imputation function}
\usage{
pipeline_llkimpute(
  data.pep.rna.mis,
  pep.ab.comp = NULL,
  nu_factor = 2,
  rna.cond.mask = NULL,
  pep.cond.mask = NULL,
  mcar = FALSE,
  protidxs = NULL,
  max.pg.size2imp = NULL,
  transpose = FALSE,
  degenerated = FALSE,
  pathifcc1 = NULL
)
}
\arguments{
\item{data.pep.rna.mis}{xxx}

\item{pep.ab.comp}{xxx}

\item{nu_factor}{xxx}

\item{rna.cond.mask}{xxx}

\item{pep.cond.mask}{xxx}

\item{mcar}{xxx}

\item{protidxs}{xxx}

\item{max.pg.size2imp}{xxx}

\item{transpose}{xxx}

\item{degenerated}{xxx}

\item{pathifcc1}{xxx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
data(bouyssie)
res.mle.transp <- pipeline_llkimpute(bouyssie)
res.mle.transp <- pipeline_llkimpute(bouyssie, transpose = TRUE)
res.mle.transp.mcar <- pipeline_llkimpute(bouyssie, transpose = TRUE, mcar = TRUE)

}
