% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pirat_config}
\alias{pirat_config}
\title{Pirat configuration information}
\usage{
pirat_config()
}
\value{
List with information on the current configuration of Pirat
}
\description{
Pirat configuration information
}
\examples{
pirat_config()

}
\keyword{internal}
