% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot2hists}
\alias{plot2hists}
\title{Plot 2 histograms}
\usage{
plot2hists(d1, d2, name1, name2, titlename, xlab = "", freq = TRUE)
}
\arguments{
\item{d1}{vector of values for the first histogram}

\item{d2}{vector of values for the first histogram}

\item{name1}{Label for first histogram}

\item{name2}{Label for 2nd histogram}

\item{titlename}{Title of figure}

\item{xlab}{X-axis label}

\item{freq}{If True, bins heights correspond to raw counts, otherwise bins 
are normalized.}
}
\value{
A plot
}
\description{
Plot 2 histograms on the same graph.
xxxxx
}
\examples{
NULL

}
