% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_pep_from_idx}
\alias{remove_pep_from_idx}
\title{xxx}
\usage{
remove_pep_from_idx(l_pep_rna, pepidxs)
}
\arguments{
\item{l_pep_rna}{xxx}

\item{pepidxs}{xxx}
}
\value{
xxx
}
\description{
Remove peptide by index
}
