% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_funcs.R
\name{split_large_pg}
\alias{split_large_pg}
\title{Split too large PGs}
\usage{
split_large_pg(adj, size_max)
}
\arguments{
\item{adj}{Adjacency matrix between peptides and PGs.}

\item{size_max}{Maximum PG size desired.}
}
\value{
New adjacency matrix between peptides and PGs.
}
\description{
Split PGs with too many peptides/precursors, while keeping other PG untouched. It creates 
new PGs with size equal to size max. Hence, some peptides can be
 duplicated in the new PGscreated. The seed is fixed such that the split generated remains constant for a given
 adjacency matrix.
}
\examples{
data(subbouyssie)
split_large_pg(subbouyssie$adj, 3)

}
