% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_large_pg_PG}
\alias{split_large_pg_PG}
\title{xxx}
\usage{
split_large_pg_PG(adj, size_max, adj_rna_pg)
}
\arguments{
\item{adj}{xxx}

\item{size_max}{xxx}

\item{adj_rna_pg}{xxx}
}
\value{
xxx
}
\description{
Split PGs that have too many peptides and transcripts.
}
\examples{
NULL

}
