% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{wrapper_pipeline_llkimpute}
\alias{wrapper_pipeline_llkimpute}
\title{xxx}
\usage{
wrapper_pipeline_llkimpute(se, ...)
}
\arguments{
\item{se}{xxx}

\item{...}{Additional arguments to pass to `pipeline_llkimpute()`}
}
\description{
xxx
}
\examples{
data(subbouyssie)
obj <- pirat2SE(subbouyssie$peptides_ab, subbouyssie$adj, subbouyssie$mask_prot_diff, 
subbouyssie$mask_pep_diff )
res <- wrapper_pipeline_llkimpute(obj)

}
