% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE.R
\name{wrapper_pipeline_llkimpute}
\alias{wrapper_pipeline_llkimpute}
\title{Imputation method using SummarizedExperiment dataset}
\usage{
wrapper_pipeline_llkimpute(se, ...)
}
\arguments{
\item{se}{An instance of the class SummarizedExperiment}

\item{...}{Additional arguments to pass to `my_pipeline_llkimpute()`}
}
\value{
See my_pipeline_llkimpute() function
}
\description{
This function imputes data from an instance of the
SummarizedExperiment structure data. After a conversion step, it calls the 
function `my_pipeline_llkimpute`.
}
\examples{
data(subbouyssie)
obj <- pirat2SE(subbouyssie$peptides_ab, subbouyssie$adj, 
subbouyssie$mask_prot_diff, subbouyssie$mask_pep_diff )
res <- wrapper_pipeline_llkimpute(obj)

}
